Workaround for bug in initialization of ffmpeg
For some reason, if the first chunk (000000)
is produced using a cv2cuda subprocess,
the subprocess is not properly terminated
when the chunk is finished, i.e. it still shows up
in the process table (as shown by ps),
and the resulting video is not readable.
Fortunately, if this chunk is made with the standard
OpenCV videowriter, and only the second one (000001) is done
with cv2cuda, then it's fine. Somehow the second chunk has no problems